requirements.txt
PyMuPDF>=1.23.7
litellm>=1.0.0
openrouter>=0.2.0
Pillow>=10.0.0
